requirements.txt
torch==1.13.0
transformers==4.24.0
pandas==1.5.2
tqdm==4.64.1
pytorch-lightning==1.8.3.post1
dataset==1.5.2
sentencepiece==0.1.97
hydra-core==1.2.0
accelerate==0.15.0
inseq==0.3.2